provider: extract fetchQuotaJSON shared helper from DeepSeek/NanoGPT quota fetchers
Phase A: the decrypt→GET→retry→decode flow was duplicated between
GetDeepSeekBalance and GetNanoGPTUsage. Extract into
fetchQuotaJSON(ctx, provider, masterKey, path, label, resource, &out)
and reduce both callers to ~6 lines. Error and log strings remain
byte-identical (label parameterized).
