+++ internal/provider/discovery_deepseek.go
@@ -4,12 +4,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/google/uuid"
 
-	"github.com/hugalafutro/model-hotel/internal/auth"
 	"github.com/hugalafutro/model-hotel/internal/debuglog"
 	"github.com/hugalafutro/model-hotel/internal/model"
 	"github.com/hugalafutro/model-hotel/internal/util"
@@ -92,38 +90,9 @@ func (d *DiscoveryService) discoverDeepSeek(ctx context.Context, provider *Provi
 
 // GetDeepSeekBalance retrieves the account balance from DeepSeek.
 func (d *DiscoveryService) GetDeepSeekBalance(ctx context.Context, provider *Provider, masterKey string) (*DeepSeekBalanceResponse, error) {
-	apiKey, err := auth.Decrypt(provider.EncryptedKey, provider.KeyNonce, provider.KeySalt, masterKey)
-	if err != nil {
-		return nil, fmt.Errorf("deepseek: failed to decrypt API key for provider %s: %w", provider.Name, err)
-	}
-
-	baseURL := util.SanitizeBaseURL(provider.BaseURL)
-	balanceURL := baseURL + "/user/balance"
-
-	req, err := http.NewRequestWithContext(ctx, "GET", balanceURL, http.NoBody)
-	if err != nil {
-		return nil, fmt.Errorf("deepseek: failed to create request for provider %s: %w", provider.Name, err)
-	}
-
-	req.Header.Set("Authorization", "Bearer "+apiKey)
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := d.doQuotaRequestWithRetry(ctx, req, provider.ID.String(), provider.Name, "deepseek")
-	if err != nil {
-		return nil, fmt.Errorf("deepseek: failed to fetch balance for provider %s: %w", provider.Name, err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		debuglog.Error("discovery: deepseek balance non-200 status", "status", resp.StatusCode, "provider", provider.Name, "provider_id", provider.ID, "body", util.SanitizeLogBody(string(body), 2000))
-		return nil, fmt.Errorf("deepseek: unexpected status code %d for provider %s", resp.StatusCode, provider.Name)
-	}
-
 	var balance DeepSeekBalanceResponse
-	if err := json.NewDecoder(resp.Body).Decode(&balance); err != nil {
-		return nil, fmt.Errorf("deepseek: failed to decode balance response for provider %s: %w", provider.Name, err)
+	if err := d.fetchQuotaJSON(ctx, provider, masterKey, "/user/balance", "deepseek", "balance", &balance); err != nil {
+		return nil, err
 	}
-
 	return &balance, nil
 }
